proof of concept for button event handler is in.  Significant work to make it go.
There is also a PIDBoom in place but no idea if it will function or not.  Needs testing and tuning.  Would replace existing algorithm.

+++ src/org/dirtymechanics/event/ButtonEventHandler.java
@@ -11,6 +11,6 @@ package org.dirtymechanics.event;
  * @author agresh
  */
 public interface ButtonEventHandler {
-    public void onEvent(long buttonEvent);
+    public void onEvent(int buttonEvent);
     
 }
